+++ CHANGELOG.rst
@@ -20,6 +20,7 @@ This version release introduces **backward incompatible changes**. Please read t
 - Added ``pokerkit.hands.RhodeIslandHoldemHand``.
 - Added ``pokerkit.lookups.RhodeIslandHoldemLookup``.
 - Added ``pokerkit.state.State.folded_status``.
+- Added ``pokerkit.state.State.total_pushed_amount``.
 
 Version 0.6.5 (November 16, 2025)
 ---------------------------------